+++ src/data_gradients/visualize/detection/__init__.py
@@ -1,0 +1,3 @@
+from data_gradients.visualize.detection.detection import draw_bboxes
+
+__all__ = ["draw_bboxes"]
