+++ src/data_gradients/visualize/detection/utils.py
@@ -1,0 +1,37 @@
+from typing import Tuple, List
+
+from matplotlib import pyplot as plt
+
+
+def best_text_color(background_color: Tuple[int, int, int]) -> Tuple[int, int, int]:
+    """Determine the best color for text to be visible on a given background color.
+
+    :param background_color: RGB values of the background color.
+    :return: RGB values of the best text color for the given background color.
+    """
+
+    # If the brightness is greater than 0.5, use black text; otherwise, use white text.
+    if compute_brightness(background_color) > 0.5:
+        return (0, 0, 0)  # Black
+    else:
+        return (255, 255, 255)  # White
+
+
+def compute_brightness(color: Tuple[int, int, int]) -> float:
+    """Computes the brightness of a given color in RGB format. From https://alienryderflex.com/hsp.html
+
+    :param color: A tuple of three integers representing the RGB values of the color.
+    :return: The brightness of the color.
+    """
+    return (0.299 * color[0] + 0.587 * color[1] + 0.114 * color[0]) / 255
+
+
+def generate_color_mapping(num_classes: int) -> List[Tuple[int, ...]]:
+    """Generate a unique BGR color for each class
+
+    :param num_classes: The number of classes in the dataset.
+    :return:            List of RGB colors for each class.
+    """
+    cmap = plt.cm.get_cmap("gist_rainbow", num_classes)
+    colors = [cmap(i, bytes=True)[:3][::-1] for i in range(num_classes)]
+    return [tuple(int(v) for v in c) for c in colors]
